lightbulb moment, we'll leave it at that ;-)
I am in open rebellion against `type *pointername`. That star means "Pointer to TYPE" in a declaration so it goes with the TYPE to tell me that I am declaring a name of a pointer, NOT the name of a variable. Dereferences go with the POINTER_NAME to get a VARIABLE VALUE so they go with the POINTER_NAME when called to Dereference it. Why is this so confusing? Why did I buy into it for so long?